% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grid_info.R
\name{add_grid_info}
\alias{add_grid_info}
\title{Add grid info}
\usage{
add_grid_info(infile, auxfile, outfile, overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{infile}{Character containing file name or path of input file.}

\item{auxfile}{Character containing file name or path of auxiliary file.}

\item{outfile}{Character containing file name or path of output file. If
NULL, the input file is directly edited instead.}

\item{overwrite}{Logical; should existing output file be overwritten? If
outfile is NULL, this parameter is ignored.}

\item{verbose}{logical; if TRUE, progress messages are shown}
}
\description{
Adds a standard longitude/latitude grid to a file which is based on a different grid.
}
\details{
No existing data is changed. The additional grid info is added as two
additional variables (lon and lat).
}
\seealso{
Other data manipulation functions: 
\code{\link{acsaf_box_mergetime}()},
\code{\link{box_mergetime}()},
\code{\link{cmsaf.transform.coordinate.system}()},
\code{\link{levbox_mergetime}()},
\code{\link{remap}()}
}
\concept{data manipulation functions}
