% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_allDatesNc.R
\name{calc_allDatesNc}
\alias{calc_allDatesNc}
\title{Designed for the CM SAF R Toolbox.}
\usage{
calc_allDatesNc(result.fileslist, ordpath)
}
\arguments{
\item{result.fileslist}{A data frame containing all meta data (data.frame).}

\item{ordpath}{NetCDF file path}
}
\description{
This function is a helper function called by the CM SAF R Toolbox.
}
