% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_overlapping_time.R
\name{calc_overlapping_time}
\alias{calc_overlapping_time}
\title{Routine to calculate overlapping time periods in two files.}
\usage{
calc_overlapping_time(var1, infile1, var2 = NULL, infile2)
}
\arguments{
\item{var1}{Name of NetCDF variable of the first data set (character).}

\item{infile1}{Filename of first input NetCDF file. This may include the directory
(character).}

\item{var2}{Name of NetCDF variable of the second data set (character).}

\item{infile2}{Filename of second input NetCDF file. This may include the directory
(character). Also supported formats for station data are .csv and .RData files.}
}
\value{
Start date and end date are the result (list).
}
\description{
Designed for CMSAF Toolbox.
}
