% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_timestepNc.R
\name{calc_timestepNc}
\alias{calc_timestepNc}
\title{Designed for the CM SAF R Toolbox.}
\usage{
calc_timestepNc(result.fileslist, ordpath)
}
\arguments{
\item{result.fileslist}{A data frame containing all meta data (data.frame).}

\item{ordpath}{NetCDF file path}
}
\description{
This function is a helper function called by the CM SAF R Toolbox.
}
