% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.coordinate.system.R
\name{check.coordinate.system}
\alias{check.coordinate.system}
\title{Designed for the CM SAF R Toolbox.}
\usage{
check.coordinate.system(nc_path, nc_temp_path, var, filelist)
}
\arguments{
\item{nc_path}{Path to NetCDF files which should be converted}

\item{nc_temp_path}{Destination NetCDF file path}

\item{var}{Name of NetCDF variable (character)}

\item{filelist}{NetCDF file names (data.frame)}
}
\description{
This function is a helper function called by the CM SAF R Toolbox.
}
