% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsafops.R
\docType{package}
\name{cmsafops-package}
\alias{cmsafops}
\alias{cmsafops-package}
\title{cmsafops: A package for analyzing and manipulating CM SAF NetCDF formatted
data.}
\description{
The 'cmsafops' functions are manipulating NetCDF input files and write the
result in a separate output file. The functions were designed and tested for
CM SAF NetCDF data, but most of the functions can be applied to other NetCDF
data, which use the CF convention and time, latitude, longitude dimensions.
As interface to NetCDF data the \link[ncdf4:ncdf4-package]{ncdf4 package} is used.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.cmsaf.eu}
  \item Report bugs at \url{https://github.com/cmsaf/cmsaf-r-tools/issues}
}

}
\author{
\strong{Maintainer}: Steffen Kothe \email{Steffen.Kothe@dwd.de}

Other contributors:
\itemize{
  \item Danny Parsons \email{danny@idems.international} [contributor]
}

}
