% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basename.R
\name{get_basename}
\alias{get_basename}
\title{Determine the basename of an NetCDF file}
\usage{
get_basename(infile, nc = NULL)
}
\arguments{
\item{infile}{Filename of input NetCDF file. This may include the directory
(character).}

\item{nc}{Alternatively to `infile` you can instead specify the input as an
object of class `ncdf4` (as returned from `ncdf4::nc_open`).}
}
\value{
A character string giving the basename
}
\description{
This function determines the basename of either a file/URL path or an 'nc'
object (using nc$filename).
}
\details{
When the origin of the file path is a local .nc
file then get_basename() is equivalent to base::basename().

get_basename() also handles the case of infile/nc being from
a URL to ensure the results always ends in ".nc". 

If both infile and nc are specified, infile is ignored.
}
