% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_time.R
\name{get_date_time}
\alias{get_date_time}
\title{Get dates and times from NetCDF type date format.}
\usage{
get_date_time(times, unit)
}
\arguments{
\item{times}{Timesteps from netcdf data (numeric).}

\item{unit}{Unit from netcdf data (character).}
}
\value{
A data frame with the columns years, months, days and times. Careful:
The parts of the date are of numeric type, but the times are stored as
characters (levels).
}
\description{
Get dates and times from NetCDF type date format.
}
\examples{
date_time <- get_date_time(times = c(159191, 5991820), 
 unit = "minutes since 1980-05-07")
date_time
date_time$years
}
