% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dimensions.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Designed for the CM SAF R Toolbox.}
\usage{
get_dimensions(id, dimnames)
}
\arguments{
\item{id}{An object of the class 'ncdf4'}

\item{dimnames}{Dimension names ('data.frame')}
}
\description{
This function is a helper function called by the CM SAF R Toolbox.
}
