% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nc_version.R
\name{get_nc_version}
\alias{get_nc_version}
\title{Designed for the CM SAF R Toolbox.}
\usage{
get_nc_version(nc34)
}
\arguments{
\item{nc34}{(numeric)}
}
\description{
This function checks the nc version.
}
