% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_processing_time_string.R
\name{get_processing_time_string}
\alias{get_processing_time_string}
\title{Get processing time string}
\usage{
get_processing_time_string(time_start, time_end)
}
\arguments{
\item{time_start}{start time of the process (of class "POSIXct" as given by
"Sys.time()")}

\item{time_end}{end time of the process (of class "POSIXct" as given by
"Sys.time()")}
}
\value{
a specialized string containing the processed time
}
\description{
Get processing time string
}
