% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_info.R
\name{get_time_info}
\alias{get_time_info}
\title{Designed for the CM SAF R Toolbox.}
\usage{
get_time_info(id, dimnames, t_name)
}
\arguments{
\item{id}{id}

\item{dimnames}{dimnames}

\item{t_name}{t_name}
}
\description{
This function is a helper function called by the CM SAF R Toolbox. Not for general use.
}
