% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Designed for the CM SAF R Toolbox.}
\usage{
read_file(infile, var_name)
}
\arguments{
\item{infile}{Filename of input NetCDF file. This may include the directory
(character).}

\item{var_name}{Name of NetCDF variable (character).}
}
\description{
This function is a helper function called by the CM SAF R Toolbox.
}
